feat(ntp): move extension field model
