+++ crafter/src/protocols/ntp/mod.rs
@@ -9,17 +9,19 @@
 
 pub mod constants;
 pub mod decode;
+pub mod extension;
 pub mod message;
 pub mod registry;
 pub mod time;
 
 pub use self::constants::*;
 pub use self::decode::decode_ntp;
+pub use self::extension::NtpExtensionField;
 pub use self::message::{
     append_ntp_packet, decode_ntp_payload, looks_like_ntp_payload, ntp_client_udp,
     ntp_ipv4_client_request, ntp_ipv6_client_request, ntp_pack_first_octet, ntp_parse_first_octet,
-    ntp_server_udp, Ntp, NtpExtensionField, NtpLeapIndicator, NtpLegacyMac, NtpMode,
-    NtpReferenceId, NtpStratum, NtpTimestamp, NtpVersion,
+    ntp_server_udp, Ntp, NtpLeapIndicator, NtpLegacyMac, NtpMode, NtpReferenceId, NtpStratum,
+    NtpTimestamp, NtpVersion,
 };
 pub use self::registry::{
     ntp_extension_field_type_meta, ntp_kiss_o_death_code_meta, ntp_leap_indicator_meta,
